+++ public/app/features/dashboard/components/DashboardRow/DashboardRow.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { DashboardRow } from '../dashgrid/DashboardRow';
-import { PanelModel } from '../panel_model';
+import { DashboardRow } from './DashboardRow';
+import { PanelModel } from '../../panel_model';
 
 describe('DashboardRow', () => {
   let wrapper, panel, dashboardMock;
