+++ public/app/features/dashboard/components/DashboardRow/index.ts
@@ -1,0 +1,1 @@
+export { DashboardRow } from './DashboardRow';
