+++ public/app/features/dashboard/index.ts
@@ -1,4 +1,4 @@
-import './dashboard_ctrl';
+import './containers/DashboardCtrl';
 import './time_srv';
 import './dashgrid/DashboardGridDirective';
