+++ public/app/routes/ReactContainer.tsx
@@ -4,8 +4,6 @@ import { Provider } from 'react-redux';
 
 import coreModule from 'app/core/core_module';
 import { store } from 'app/store/store';
-import { BackendSrv } from 'app/core/services/backend_srv';
-import { DatasourceSrv } from 'app/features/plugins/datasource_srv';
 import { ContextSrv } from 'app/core/services/context_srv';
 
 function WrapInProvider(store, Component, props) {
@@ -20,8 +18,6 @@ function WrapInProvider(store, Component, props) {
 export function reactContainer(
   $route,
   $location,
-  backendSrv: BackendSrv,
-  datasourceSrv: DatasourceSrv,
   contextSrv: ContextSrv
 ) {
   return {
@@ -42,11 +38,7 @@ export function reactContainer(
         component = component.default;
       }
 
-      const props = {
-        backendSrv: backendSrv,
-        datasourceSrv: datasourceSrv,
-        routeParams: $route.current.params,
-      };
+      const props = { };
 
       ReactDOM.render(WrapInProvider(store, component, props), elem[0]);
