+++ public/sass/components/_view_states.scss
@@ -50,3 +50,4 @@
     display: none;
   }
 }
+
